- Fixed count in kv store  - Cleaned up system prompt  - Added name back into messages in chat

+++ Mute.Moe/Services/Database/IKeyValueStorage.cs
@@ -18,7 +18,7 @@ public interface IKeyValueStorage
     /// Count the total number of items
     /// </summary>
     /// <returns></returns>
-    public Task<int> Count();
+    public Task<long> Count();
 
     /// <summary>
     /// Delete all items
